- Added compatibility with v1.12.0 (including E3 Definitive Mod support!) - Added "Renderer CVars list" (Misc) for changing game rendering variables (only float for now) - Changed DirectX 11 and DirectX 12 hooks to be more reliable and less likely to crash the game (removed kiero dependency) - Fixed time and weather changing taking a long time to apply (now it is instant!) - Fixed ownership of Player Variables managed by EGameTools, removing conflict in managing Player Variables between different modules (e.g. Player and World) and any bugs associated with this issue - Fixed conflict between ImGui state and Player Variables list, causing the game to crash when changing Player Variables - Removed Windows.h dependency where possible - General code cleanup - Added a way of managing dynamic offsets based on game version - Added a way to stop hooks from running on startup, requiring manual hooking

+++ EGameSDK/include/EGSDK/Utils/Time.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 #include <chrono>
 #include <iomanip>
 #include <EGSDK\Exports.h>
@@ -8,16 +9,16 @@ namespace EGSDK::Utils {
 		class EGameSDK_API Timer {
 			using clock = std::chrono::steady_clock;
 		public:
-			long long timeToPass;
+			int64_t timeToPass;
 			Timer() : timeToPass(0), timePassed(false), start(0), end(0) {}
-			Timer(long timeMs);
+			Timer(int64_t timeMs);
 
-			const long long GetTimePassed() const;
+			const int64_t GetTimePassed() const;
 			const bool DidTimePass();
 			void Reset();
 		private:
-			long long start;
-			long long end;
+			int64_t start;
+			int64_t end;
 			bool timePassed;
 		};
